requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==6.5.0
cssbeautifier==1.14.7
Django==2.2.16
django-aggregates==0.1.3
django-debug-toolbar==3.2.4
djhtml==2.0.0
EditorConfig==0.12.3
exceptiongroup==1.1.0
Faker==12.0.1
flake8==5.0.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
jsbeautifier==1.14.7
mccabe==0.7.0
mixer==7.1.2
packaging==21.3
pathspec==0.9.0
Pillow==8.3.1
pluggy==0.13.1
psycopg2-binary==2.9.6
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.7
pytest==6.2.4
pytest-cov==3.0.0
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.1
PyYAML==6.0
regex==2022.10.31
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.9
zipp==3.7.0
djangorestframework==3.12.4
